refactor: valid uuid back-end / post contaccts / refactor http client

+++ back-end/src/app/controllers/ContactController.js
@@ -1,4 +1,6 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID')
+
 class ContactController {
     
     async index(request, response) {
@@ -10,6 +12,10 @@ class ContactController {
 
     async show(request, response) {
         const { id } = request.params;
+
+        if (!isValidUUID(id)) {
+            return response.status(400).json({ error: 'Invalid contact ID' });
+        }
 
         const contact = await ContactsRepository.findById(id);
 
@@ -27,13 +33,23 @@ class ContactController {
             return response.status(400).json({ error: 'Name is required' });
         }
 
-        const contactExist = await ContactsRepository.findByEmail(email);
-        if(contactExist){
-            return response.status(400).json({ error: 'This e-mail is aleardy been taken' });
+        if (category_id && !isValidUUID(category_id)) {
+            return response.status(400).json({ error: 'Invalid category ID' });
+        }
+        
+        if (email) {
+            const contactExist = await ContactsRepository.findByEmail(email);
+            
+            if(contactExist){
+                return response.status(400).json({ error: 'This e-mail is aleardy been taken' });
+            }
         }
 
         const contact = await ContactsRepository.create({
-            name, email, phone, category_id
+            name, 
+            email: email || null, 
+            phone, 
+            category_id: category_id || null
         });
 
         response.status(201).json(contact); 
@@ -43,22 +59,36 @@ class ContactController {
         const { id } = request.params;
         const { name, email, phone, category_id } = request.body;
 
-        const contactExist = await ContactsRepository.findById(id);
-        if(!contactExist){
-            return response.status(404).json({ error: 'User not found' });
+        if (!isValidUUID(id)) {
+            return response.status(400).json({ error: 'Invalid contact ID' });
+        }
+
+        if (category_id && !isValidUUID(category_id)) {
+            return response.status(400).json({ error: 'Invalid category ID' });
         }
 
         if(!name){
             return response.status(400).json({ error: 'Name is required' });
         }
 
-        const contactByEmail = await ContactsRepository.findByEmail(email);
-        if(contactByEmail && contactByEmail.id != id){
-            return response.status(400).json({ error: 'This e-mail is aleardy in use' });
+        const contactExist = await ContactsRepository.findById(id);
+        if(!contactExist){
+            return response.status(404).json({ error: 'User not found' });
+        }
+
+        if (email) {
+            const contactByEmail = await ContactsRepository.findByEmail(email);
+            
+            if(contactByEmail && contactByEmail.id != id){
+                return response.status(400).json({ error: 'This e-mail is aleardy in use' });
+            }
         }
 
         const contact = await ContactsRepository.update(id, {
-            name, email, phone, category_id
+            name, 
+            email: email || null, 
+            phone, 
+            category_id: category_id || null
         });
 
         response.json(contact);
@@ -66,6 +96,10 @@ class ContactController {
 
     async delete(request, response) {
         const { id } = request.params;
+
+        if (!isValidUUID(id)) {
+            return response.status(400).json({ error: 'Invalid contact ID' });
+        }
 
         await ContactsRepository.delete(id);
         response.sendStatus(204);
